Formatting

+++ CvGameCoreDLL_Expansion2/CvTeam.cpp
@@ -4325,34 +4325,31 @@ void CvTeam::setAtWar(TeamTypes eIndex, bool bNewValue, bool bAggressorPacifier)
 	CvAssertMsg(eIndex >= 0, "eIndex is expected to be non-negative (invalid Index)");
 	CvAssertMsg(eIndex < MAX_TEAMS, "eIndex is expected to be within maximum bounds (invalid Index)");
 	CvAssertMsg(eIndex != GetID() || bNewValue == false, "Team is setting war with itself!");
-	if (eIndex != GetID() || bNewValue == false)
+	if (eIndex != GetID() || !bNewValue)
 	{
 		m_abAggressorPacifier[eIndex] = bAggressorPacifier;
 		m_abAtWar[eIndex] = bNewValue;
-		if (MOD_BALANCE_CORE)
+		for (int iAttackingPlayer = 0; iAttackingPlayer < MAX_MAJOR_CIVS; iAttackingPlayer++)
 		{
-			for (int iAttackingPlayer = 0; iAttackingPlayer < MAX_MAJOR_CIVS; iAttackingPlayer++)
+			PlayerTypes eAttackingPlayer = (PlayerTypes)iAttackingPlayer;
+			CvPlayerAI& kAttackingPlayer = GET_PLAYER(eAttackingPlayer);
+			if (kAttackingPlayer.isAlive() && kAttackingPlayer.getTeam() == GetID())
 			{
-				PlayerTypes eAttackingPlayer = (PlayerTypes)iAttackingPlayer;
-				CvPlayerAI& kAttackingPlayer = GET_PLAYER(eAttackingPlayer);
-				if (kAttackingPlayer.isAlive() && kAttackingPlayer.getTeam() == GetID())
+				for (int iDefendingPlayer = 0; iDefendingPlayer < MAX_MAJOR_CIVS; iDefendingPlayer++)
 				{
-					for (int iDefendingPlayer = 0; iDefendingPlayer < MAX_MAJOR_CIVS; iDefendingPlayer++)
+					PlayerTypes eDefendingPlayer = (PlayerTypes)iDefendingPlayer;
+					CvPlayerAI& kDefendingPlayer = GET_PLAYER(eDefendingPlayer);
+					if (kDefendingPlayer.isAlive() && kDefendingPlayer.getTeam() == eIndex)
 					{
-						PlayerTypes eDefendingPlayer = (PlayerTypes)iDefendingPlayer;
-						CvPlayerAI& kDefendingPlayer = GET_PLAYER(eDefendingPlayer);
-						if (kDefendingPlayer.isAlive() && kDefendingPlayer.getTeam() == eIndex)
-						{
-							kAttackingPlayer.recomputeGreatPeopleModifiers();
-							kDefendingPlayer.recomputeGreatPeopleModifiers();
-						}
+						kAttackingPlayer.recomputeGreatPeopleModifiers();
+						kDefendingPlayer.recomputeGreatPeopleModifiers();
 					}
 				}
 			}
 		}
 	}
 
-	if (MOD_BALANCE_CORE && bNewValue)
+	if (bNewValue)
 	{
 		//Check for bad units, and capture them!
 		vector<CvUnitCaptureDefinition> kCaptureUnitList;
@@ -4367,45 +4364,42 @@ void CvTeam::setAtWar(TeamTypes eIndex, bool bNewValue, bool bAggressorPacifier)
 				CvUnit* pLoopUnit = NULL; //for some stupid reason createCaptureUnit is a member of CvUnit and not CvPlayer
 				for (pLoopUnit = kPlayer.firstUnit(&iLoop); pLoopUnit; pLoopUnit = kPlayer.nextUnit(&iLoop))
 				{
-					if (pLoopUnit->IsCombatUnit())
+					if (pLoopUnit->IsCombatUnit() && pLoopUnit->plot() != NULL)
 					{
-						if (pLoopUnit->plot() != NULL)
+						for (int iUnitLoop = 0; iUnitLoop < pLoopUnit->plot()->getNumUnits(); iUnitLoop++)
 						{
-							for (int iUnitLoop = 0; iUnitLoop < pLoopUnit->plot()->getNumUnits(); iUnitLoop++)
+							CvUnit* pPotentialCaptureUnit = pLoopUnit->plot()->getUnitByIndex(iUnitLoop);
+
+							if (pPotentialCaptureUnit &&
+								pPotentialCaptureUnit->getTeam() == eIndex && //only from the right team
+								!pPotentialCaptureUnit->IsCombatUnit() &&
+								!pPotentialCaptureUnit->isEmbarked() &&
+								!pPotentialCaptureUnit->isDelayedDeath()) //can only capture once!
 							{
-								CvUnit* pPotentialCaptureUnit = pLoopUnit->plot()->getUnitByIndex(iUnitLoop);
-
-								if (pPotentialCaptureUnit &&
-									pPotentialCaptureUnit->getTeam() == eIndex && //only from the right team
-									!pPotentialCaptureUnit->IsCombatUnit() &&
-									!pPotentialCaptureUnit->isEmbarked() &&
-									!pPotentialCaptureUnit->isDelayedDeath()) //can only capture once!
+								if (pPotentialCaptureUnit->getCaptureUnitType(GET_PLAYER(pPotentialCaptureUnit->getOwner()).getCivilizationType()) != NO_UNIT)
 								{
-									if (pPotentialCaptureUnit->getCaptureUnitType(GET_PLAYER(pPotentialCaptureUnit->getOwner()).getCivilizationType()) != NO_UNIT)
+									CvUnitCaptureDefinition kCaptureDef;
+									if (pPotentialCaptureUnit->getCaptureDefinition(&kCaptureDef, kPlayer.GetID()))
 									{
-										CvUnitCaptureDefinition kCaptureDef;
-										if (pPotentialCaptureUnit->getCaptureDefinition(&kCaptureDef, kPlayer.GetID()))
+										bool bAlreadyCaptured = false;
+										for (uint uiCaptureIndex = 0; uiCaptureIndex < kCaptureUnitList.size(); ++uiCaptureIndex)
 										{
-											bool bAlreadyCaptured = false;
-											for (uint uiCaptureIndex = 0; uiCaptureIndex < kCaptureUnitList.size(); ++uiCaptureIndex)
+											if (kCaptureUnitList[uiCaptureIndex].iUnitID == kCaptureDef.iUnitID)
 											{
-												if (kCaptureUnitList[uiCaptureIndex].iUnitID == kCaptureDef.iUnitID)
-												{
-													bAlreadyCaptured = true;
-													break;
-												}
+												bAlreadyCaptured = true;
+												break;
 											}
-											if (!bAlreadyCaptured)
-											{
-												kCaptureUnitList.push_back(kCaptureDef);
-											}
-											pPotentialCaptureUnit->setCapturingPlayer(NO_PLAYER);	// Make absolutely sure this is not valid so the kill does not do the capture.
 										}
+										if (!bAlreadyCaptured)
+										{
+											kCaptureUnitList.push_back(kCaptureDef);
+										}
+										pPotentialCaptureUnit->setCapturingPlayer(NO_PLAYER);	// Make absolutely sure this is not valid so the kill does not do the capture.
 									}
-
-									//be careful here, it's possible we're about to kill a civilian which is right now executing a mission causing this war state change
-									pPotentialCaptureUnit->kill(true, kPlayer.GetID());
 								}
+
+								//be careful here, it's possible we're about to kill a civilian which is right now executing a mission causing this war state change
+								pPotentialCaptureUnit->kill(true, kPlayer.GetID());
 							}
 						}
 					}
